requirements.txt
fastapi>=0.111.1
uvicorn[standard]>=0.23.1
python-dotenv>=1.0.1
openai>=1.0.0
requests>=2.31.0
supabase>=2.7.1
